Update filter component
