+++ apps/web/@/utils/capitalize.ts
@@ -1,0 +1,3 @@
+export const capitalizeFirstLetter = (value: string) => {
+  return value.charAt(0).toUpperCase() + value.slice(1)
+}
